Move compute benchmark utils in header
